refactor(config): move tracing related config into own file
